pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    { name = "y" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
    "cffi (==2.0.0b1)",
    "torch (==2.8.0)",
    "langchain (==0.3.27)",
    "aiofiles (==24.1.0)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.12.15)",
    "aiosignal (==1.4.0)",
    "annotated-types (==0.7.0)",
    "anyio (==4.10.0)",
    "argon2-cffi (==25.1.0)",
    "argon2-cffi-bindings (==25.1.0)",
    "arrow (==1.3.0)",
    "asttokens (==3.0.0)",
    "async-lru (==2.0.5)",
    "attrs (==25.3.0)",
    "babel (==2.17.0)",
    "backoff (==2.2.1)",
    "bcrypt (==4.3.0)",
    "beautifulsoup4 (==4.13.5)",
    "bleach (==6.2.0)",
    "build (==1.3.0)",
    "cachetools (==5.5.2)",
    "certifi (==2025.8.3)",
    "charset-normalizer (==3.4.3)",
    "chromadb (==1.0.20)",
    "click (==8.2.1)",
    "colorama (==0.4.6)",
    "coloredlogs (==15.0.1)",
    "comm (==0.2.3)",
    "cryptography (==45.0.6)",
    "dataclasses-json (==0.6.7)",
    "debugpy (==1.8.16)",
    "decorator (==5.2.1)",
    "defusedxml (==0.7.1)",
    "distro (==1.9.0)",
    "durationpy (==0.10)",
    "emoji (==2.14.1)",
    "executing (==2.2.0)",
    "fastjsonschema (==2.21.2)",
    "filelock (==3.19.1)",
    "filetype (==1.2.0)",
    "flatbuffers (==25.2.10)",
    "fqdn (==1.5.1)",
    "frozenlist (==1.7.0)",
    "fsspec (==2025.7.0)",
    "google-auth (==2.40.3)",
    "googleapis-common-protos (==1.70.0)",
    "greenlet (==3.2.4)",
    "groq (==0.31.0)",
    "grpcio (==1.74.0)",
    "h11 (==0.16.0)",
    "html5lib (==1.1)",
    "httpcore (==1.0.9)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "httpx-sse (==0.4.1)",
    "huggingface-hub (==0.34.4)",
    "humanfriendly (==10.0)",
    "idna (==3.10)",
    "importlib-metadata (==8.7.0)",
    "importlib-resources (==6.5.2)",
    "ipykernel (==6.30.1)",
    "ipython (==9.5.0)",
    "ipython-pygments-lexers (==1.1.1)",
    "ipywidgets (==8.1.7)",
    "isoduration (==20.11.0)",
    "jedi (==0.19.2)",
    "jinja2 (==3.1.6)",
    "jiter (==0.10.0)",
    "joblib (==1.5.2)",
    "json5 (==0.12.1)",
    "jsonpatch (==1.33)",
    "jsonpointer (==3.0.0)",
    "jsonschema (==4.25.1)",
    "jsonschema-specifications (==2025.4.1)",
    "jupyter (==1.1.1)",
    "jupyter-console (==6.6.3)",
    "jupyter-events (==0.12.0)",
    "jupyter-lsp (==2.3.0)",
    "jupyter-client (==8.6.3)",
    "jupyter-core (==5.8.1)",
    "jupyter-server (==2.17.0)",
    "jupyter-server-terminals (==0.5.3)",
    "jupyterlab (==4.4.6)",
    "jupyterlab-pygments (==0.3.0)",
    "jupyterlab-server (==2.27.3)",
    "jupyterlab-widgets (==3.0.15)",
    "kubernetes (==33.1.0)",
    "langchain-community (==0.3.29)",
    "langchain-core (==0.3.75)",
    "langchain-groq (==0.3.7)",
    "langchain-huggingface (==0.3.1)",
    "langchain-openai (==0.3.32)",
    "langchain-text-splitters (==0.3.10)",
    "langchainhub (==0.1.21)",
    "langdetect (==1.0.9)",
    "langsmith (==0.4.21)",
    "lark (==1.2.2)",
    "lxml (==6.0.1)",
    "markdown-it-py (==4.0.0)",
    "markupsafe (==3.0.2)",
    "marshmallow (==3.26.1)",
    "matplotlib-inline (==0.1.7)",
    "mdurl (==0.1.2)",
    "mistune (==3.1.4)",
    "mmh3 (==5.2.0)",
    "mpmath (==1.3.0)",
    "multidict (==6.6.4)",
    "mypy-extensions (==1.1.0)",
    "nbclient (==0.10.2)",
    "nbconvert (==7.16.6)",
    "nbformat (==5.10.4)",
    "nest-asyncio (==1.6.0)",
    "networkx (==3.5)",
    "nltk (==3.9.1)",
    "notebook (==7.4.5)",
    "notebook-shim (==0.2.4)",
    "numpy (==2.3.2)",
    "oauthlib (==3.3.1)",
    "olefile (==0.47)",
    "onnxruntime (==1.22.1)",
    "openai (==1.102.0)",
    "opentelemetry-api (==1.36.0)",
    "opentelemetry-exporter-otlp-proto-common (==1.36.0)",
    "opentelemetry-exporter-otlp-proto-grpc (==1.36.0)",
    "opentelemetry-proto (==1.36.0)",
    "opentelemetry-sdk (==1.36.0)",
    "opentelemetry-semantic-conventions (==0.57b0)",
    "orjson (==3.11.3)",
    "overrides (==7.7.0)",
    "packaging (==24.2)",
    "pandocfilters (==1.5.1)",
    "parso (==0.8.5)",
    "pillow (==11.3.0)",
    "platformdirs (==4.4.0)",
    "playwright (==1.55.0)",
    "posthog (==5.4.0)",
    "prometheus-client (==0.22.1)",
    "prompt-toolkit (==3.0.52)",
    "propcache (==0.3.2)",
    "protobuf (==6.32.0)",
    "psutil (==7.0.0)",
    "pure-eval (==0.2.3)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pybase64 (==1.4.2)",
    "pycparser (==2.22)",
    "pydantic (==2.11.7)",
    "pydantic-settings (==2.10.1)",
    "pydantic-core (==2.33.2)",
    "pyee (==13.0.0)",
    "pygments (==2.19.2)",
    "pypdf (==6.0.0)",
    "pypika (==0.48.9)",
    "pyproject-hooks (==1.2.0)",
    "pyreadline3 (==3.5.4)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.1.1)",
    "python-iso639 (==2025.2.18)",
    "python-json-logger (==3.3.0)",
    "python-magic (==0.4.27)",
    "python-oxmsg (==0.0.2)",
    "pywin32 (==311)",
    "pywinpty (==3.0.0)",
    "pyyaml (==6.0.2)",
    "pyzmq (==27.0.2)",
    "rapidfuzz (==3.14.0)",
    "referencing (==0.36.2)",
    "regex (==2025.8.29)",
    "requests (==2.32.5)",
    "requests-oauthlib (==2.0.0)",
    "requests-toolbelt (==1.0.0)",
    "rfc3339-validator (==0.1.4)",
    "rfc3986-validator (==0.1.1)",
    "rfc3987-syntax (==1.1.0)",
    "rich (==14.1.0)",
    "rpds-py (==0.27.1)",
    "rsa (==4.9.1)",
    "safetensors (==0.6.2)",
    "scikit-learn (==1.7.1)",
    "scipy (==1.16.1)",
    "send2trash (==1.8.3)",
    "sentence-transformers (==5.1.0)",
    "setuptools (==80.9.0)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "soupsieve (==2.8)",
    "sqlalchemy (==2.0.43)",
    "stack-data (==0.6.3)",
    "sympy (==1.14.0)",
    "tenacity (==9.1.2)",
    "terminado (==0.18.1)",
    "threadpoolctl (==3.6.0)",
    "tiktoken (==0.11.0)",
    "tinycss2 (==1.4.0)",
    "tokenizers (==0.22.0)",
    "tornado (==6.5.2)",
    "tqdm (==4.67.1)",
    "traitlets (==5.14.3)",
    "transformers (==4.56.0)",
    "typer (==0.16.1)",
    "types-python-dateutil (==2.9.0.20250822)",
    "types-requests (==2.32.4.20250809)",
    "typing-inspect (==0.9.0)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.15.0)",
    "unstructured (==0.18.14)",
    "unstructured-client (==0.42.3)",
    "uri-template (==1.3.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.35.0)",
    "watchfiles (==1.1.0)",
    "wcwidth (==0.2.13)",
    "webcolors (==24.11.1)",
    "webencodings (==0.5.1)",
    "websocket-client (==1.8.0)",
    "websockets (==15.0.1)",
    "widgetsnbextension (==4.0.14)",
    "wrapt (==1.17.3)",
    "yarl (==1.20.1)",
    "zipp (==3.23.0)",
    "zstandard (==0.24.0)",
    "fastapi (>=0.116.1,<0.117.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
